+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
 - ``enums.PaymentMethodType`` has been deprecated, use ``enums.DjstripePaymentMethodType``
 - Made ``SubscriptionItem.quantity`` nullable as per Plans with ``usage_type="metered"`` (follow-up to #865)
 - Added manage command ``djstripe_sync_models`` (#727, #89)
+- Fixed issue with re-creating a customer after `Customer.purge()` (#916)
 
 Changes from API 2018-11-08: